BP_Gate_Valve has been added, Works for multiple functions, see change description
This BP handles everything a door needs in a game. Default instance editable variables allow level designers to easily place gates, valves, and regular doors all with the same blueprint. simply check the boolean boxes for desired result. secondly a construction script changes the location and shape of the valve trigger volume so a designer may change that and see their changes instantly in editor.